mptcp: olia: Fix possible integer overflow
Multiplying two 32-bit integers (even if they are stored in a 64-bit
int) will be doing 32-bit arithmetic and only then will be cast to
64-bit.
So, we need to explicitly cast beforehand.

Fixes: 514e1d11669c9 (Initial OLIA-implementation)

+++ net/mptcp/mptcp_olia.c
@@ -167,7 +167,7 @@ static void mptcp_get_epsilon(struct mptcp_cb *mpcb)
 			ca->epsilon_num = 0;
 			ca->epsilon_den = 1;
 		} else {
-			tmp_rtt = tp->srtt * tp->srtt;
+			tmp_rtt = (u64)tp->srtt * tp->srtt;
 			tmp_int = max(ca->mptcp_loss3 - ca->mptcp_loss2,
 				      ca->mptcp_loss2 - ca->mptcp_loss1);
 			tmp_cwnd = mptcp_get_crt_cwnd(sk);
